+++ conduit/data/datamodules/vision/celeba.py
@@ -8,11 +8,10 @@ from kit.torch import TrainingMode, prop_random_split
 from pytorch_lightning import LightningDataModule
 
 from conduit.data.datamodules.base import CdtDataModule
+from conduit.data.datamodules.vision.base import CdtVisionDataModule
 from conduit.data.datasets.utils import ImageTform
 from conduit.data.datasets.vision.celeba import CelebA, CelebASplit, CelebAttr
 from conduit.data.structures import TrainValTestSplit
-
-from .base import CdtVisionDataModule
 
 __all__ = ["CelebADataModule"]
