+++ conduit/data/datamodules/vision/cmnist.py
@@ -12,11 +12,10 @@ from pytorch_lightning import LightningDataModule
 from torchvision.datasets import MNIST
 
 from conduit.data.datamodules.base import CdtDataModule
+from conduit.data.datamodules.vision.base import CdtVisionDataModule
 from conduit.data.datasets.utils import ImageTform
 from conduit.data.datasets.vision.cmnist import ColoredMNIST, ColoredMNISTSplit
 from conduit.data.structures import MeanStd, TrainValTestSplit
-
-from .base import CdtVisionDataModule
 
 __all__ = ["ColoredMNISTDataModule"]
