+++ conduit/data/datasets/audio/ecoacoustics.py
@@ -32,9 +32,8 @@ from torchvision.datasets.utils import (
 from tqdm import tqdm
 from typing_extensions import Literal
 
+from conduit.data.datasets.audio.base import CdtAudioDataset
 from conduit.data.datasets.utils import AudioTform, FileInfo
-
-from .base import CdtAudioDataset
 
 __all__ = ["Ecoacoustics"]
