+++ conduit/data/datasets/vision/celeba.py
@@ -11,9 +11,9 @@ import torch
 
 from conduit.data.datasets.utils import FileInfo, ImageTform, download_from_gdrive
 
-from .base import CdtVisionDataset
-
 __all__ = ["CelebA", "CelebAttr", "CelebASplit"]
+
+from conduit.data.datasets.vision.base import CdtVisionDataset
 
 
 class CelebAttr(Enum):
