+++ conduit/data/datasets/vision/cmnist.py
@@ -17,12 +17,12 @@ from typing_extensions import Literal
 
 from conduit.data.datasets.utils import ImageTform, RawImage
 
-from .base import CdtVisionDataset
-
 __all__ = [
     "ColoredMNIST",
     "ColoredMNISTSplit",
 ]
+
+from conduit.data.datasets.vision.base import CdtVisionDataset
 
 
 @overload
