+++ conduit/data/datasets/vision/ssrp.py
@@ -10,9 +10,9 @@ import torch
 
 from conduit.data.datasets.utils import FileInfo, ImageTform, download_from_gdrive
 
-from .base import CdtVisionDataset
-
 __all__ = ["SSRP", "SSRPSplit"]
+
+from conduit.data.datasets.vision.base import CdtVisionDataset
 
 
 class SSRPSplit(Enum):
