+++ conduit/fair/data/datamodules/tabular/admissions.py
@@ -7,9 +7,9 @@ from ethicml.preprocessing.scaling import ScalerType
 from kit import parsable
 from kit.torch import TrainingMode
 
-from .base import EthicMlDataModule
-
 __all__ = ["AdmissionsDataModule"]
+
+from conduit.fair.data.datamodules.tabular.base import EthicMlDataModule
 
 
 class AdmissionsSens(Enum):
