+++ conduit/fair/data/datamodules/tabular/adult.py
@@ -7,9 +7,9 @@ from ethicml.preprocessing.scaling import ScalerType
 from kit import parsable
 from kit.torch import TrainingMode
 
-from .base import EthicMlDataModule
-
 __all__ = ["AdultDataModule"]
+
+from conduit.fair.data.datamodules.tabular.base import EthicMlDataModule
 
 
 class AdultSens(Enum):
