+++ conduit/models/erm.py
@@ -11,16 +11,15 @@ import torch
 from torch import Tensor, nn
 
 from conduit.data import BinarySample
-from conduit.types import Loss, MetricDict, Stage
-
-from .base import CdtModel
-from .utils import (
+from conduit.models.base import CdtModel
+from conduit.models.utils import (
     accuracy,
     aggregate_over_epoch,
     make_no_grad,
     precision_at_k,
     prefix_keys,
 )
+from conduit.types import Loss, MetricDict, Stage
 
 __all__ = ["ERMClassifier", "FineTuner"]
