+++ conduit/models/self_supervised/moco/transforms.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
+import random
 from typing import Sequence
 
+from PIL import Image, ImageFilter
 from torchvision import transforms as T
 from torchvision.transforms.functional import InterpolationMode
 
@@ -8,13 +10,27 @@ from conduit.constants import IMAGENET_STATS
 from conduit.data.datasets.utils import PillowTform
 from conduit.data.structures import MeanStd
 
-__all__ = [
-    "moco_ft_transform",
-    "moco_test_transform",
-    "mocov2_train_transform",
-]
+__all__ = ["moco_ft_transform", "moco_test_transform", "mocov2_train_transform", "GaussianBlur"]
 
-from conduit.models.self_supervised.transforms import GaussianBlur
+
+class GaussianBlur:
+    """
+    Apply Gaussian Blur to the PIL image with some probability.
+    """
+
+    def __init__(self, p: float = 0.5, *, radius_min: float = 0.1, radius_max: float = 2.0) -> None:
+        self.prob = p
+        self.radius_min = radius_min
+        self.radius_max = radius_max
+
+    def __call__(self, img: Image.Image) -> Image.Image:
+        do_it = random.random() <= self.prob
+        if not do_it:
+            return img
+
+        return img.filter(
+            ImageFilter.GaussianBlur(radius=random.uniform(self.radius_min, self.radius_max))
+        )
 
 
 def mocov2_train_transform(
