--- conduit/models/self_supervised/transforms.py
@@ -1,25 +1,0 @@
-import random
-
-from PIL import Image, ImageFilter
-
-__all__ = ["GaussianBlur"]
-
-
-class GaussianBlur:
-    """
-    Apply Gaussian Blur to the PIL image with some probability.
-    """
-
-    def __init__(self, p: float = 0.5, *, radius_min: float = 0.1, radius_max: float = 2.0) -> None:
-        self.prob = p
-        self.radius_min = radius_min
-        self.radius_max = radius_max
-
-    def __call__(self, img: Image.Image) -> Image.Image:
-        do_it = random.random() <= self.prob
-        if not do_it:
-            return img
-
-        return img.filter(
-            ImageFilter.GaussianBlur(radius=random.uniform(self.radius_min, self.radius_max))
-        )
